Moving createSearchClause to utils and adding tag searching
