+++ config.yaml
@@ -1,1 +1,1 @@
-endpoint: https://recipe-palette.herokuapp.com/
+endpoint: https://recipe-palette-dev.herokuapp.com/
