+++ src/templates/search.js
@@ -4,6 +4,7 @@ import queryString from 'query-string'
 import { useQuery } from '@apollo/react-hooks'
 import gql from 'graphql-tag'
 import { useAuth } from 'react-use-auth'
+import { isEmpty } from 'lodash'
 
 import Title from '../components/title'
 import {
@@ -13,6 +14,7 @@ import {
 import CardGrid from '../components/card-grid'
 import { RecipeCardGridLoader } from '../components/recipe-card-loader'
 import SearchBar from '../components/search-bar'
+import { createSearchClause } from '../utils/createSearchClause'
 
 const SEARCH_QUERY = gql`
   query SearchQuery($whereClause: recipe_bool_exp!, $user_id: String!) {
@@ -27,40 +29,13 @@ const SEARCH_QUERY = gql`
   ${recipeCardInformationFragment}
 `
 
-// Function to dynamically create where clause based on query string inputs
-const createWhereClase = parsedSearch => {
-  const { q, ingredients } = parsedSearch
-  const whereClause = {}
-  if (q) {
-    whereClause.latest = {
-      name: {
-        _ilike: `%${q}%`,
-      },
-    }
-  }
-
-  if (ingredients) {
-    let ingredientQuery = ingredients
-    if (typeof ingredients === 'object') {
-      ingredientQuery = ingredients.join('|')
-    }
-    whereClause.ingredients = {
-      name: {
-        _similar: `%(${ingredientQuery})%`,
-      },
-    }
-  }
-
-  return whereClause
-}
-
 const Search = ({ location }) => {
   const { userId } = useAuth()
   const parsedSearch = queryString.parse(location.search, {
     arrayFormat: 'comma',
   })
   const { q } = parsedSearch
-  const whereClause = createWhereClase(parsedSearch)
+  const whereClause = createSearchClause(parsedSearch)
 
   const { data: searchData, loading } = useQuery(SEARCH_QUERY, {
     variables: {
@@ -69,16 +44,9 @@ const Search = ({ location }) => {
     },
   })
 
-  return q ? (
-    <div sx={{ py: `4` }}>
-      <Title>Search results for {q}</Title>
-      {loading ? (
-        <RecipeCardGridLoader />
-      ) : (
-        <CardGrid recipes={searchData.recipes} />
-      )}
-    </div>
-  ) : (
+  console.log(parsedSearch)
+
+  return isEmpty(parsedSearch) ? (
     <div sx={{ py: `4` }}>
       <Title>Search for Recipes</Title>
       <p>
@@ -87,6 +55,15 @@ const Search = ({ location }) => {
         bookmark, and edit.
       </p>
       <SearchBar />
+    </div>
+  ) : (
+    <div sx={{ py: `4` }}>
+      <Title>Search results for {q}</Title>
+      {loading ? (
+        <RecipeCardGridLoader />
+      ) : (
+        <CardGrid recipes={searchData.recipes} />
+      )}
     </div>
   )
 }
